fix: filtered issues in archived repositories
