--- .github/workflows/test.yml
@@ -1,23 +1,0 @@
-name: test
-
-on:
-  pull_request:
-    types:
-      - "opened"
-
-permissions:
-  contents: write
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: 1.18
-
-      - name: Test
-        run: go test -v ./...
