+++ data/repository.go
@@ -1,0 +1,7 @@
+package data
+
+type Repository struct {
+	Name          string
+	NameWithOwner string
+	IsArchived    bool
+}
